about us video and responsiveness
re-worked the embedded video to add an overlay that disappears when playing and reappears with a pause video when clicked. used jquery javascript to do this -- fadein and fadeout for the overlay itself and removeClass and addClass for the sitch from play to pause icons. added a video section in place of the video header for mobile devices. removed the extra sections from about-us.